pip-package/pyproject.toml
# pyproject.toml

[project]
name = "substrate-ai"
# dynamic = ["version"]
version = "0.0.4"
description = "Substrate AI package"
readme = "README.md"
authors = [{ name = "SubstrateAI", email = "support@substrateai.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["cloud", "gpu", "ml", "ai"]

dependencies = [
    "boto3 >= 1.28.0",
    "docker >= 6.1.0",
    "typer[all] >= 0.9.0",
    "human-id >= 0.2.0",
    "python-on-whales == 0.64.2",
    "yaspin == 3.0.0",
    "supabase == 1.0.4",
    "google-cloud-artifact-registry == 1.8.3",
    "google-cloud-container == 2.30.0",
    "envd == 0.3.40",
    "pyyaml == 6.0.1"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://docs.substrateai.com/"

[project.scripts]
substrate-ai = "cli:app"

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.bumpver]
current_version = "0.0.4"
version_pattern = 'MAJOR.MINOR.PATCH'
# commit_message = "bump version {old_version} -> {new_version}"
# tag_message = "{new_version}"
# tag_scope = "default"
# commit = true
# tag = true
# push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"src/__init__.py" = ["{version}"]


[tool.ruff]
line-length = 200
select = ["E", "F", "N"]

